--- a/src/benchmarks/micro/sve/StrIndexOf.cs
+++ b/src/benchmarks/micro/sve/StrIndexOf.cs
@@ -26,7 +26,7 @@ namespace SveBenchmarks
             }
         }
 
-        [Params(15, 127, 527, 10015)]
+        [Params(19, 127, 527, 10015)]
         public int Size;
 
         private char[] _array;
@@ -42,16 +42,19 @@ namespace SveBenchmarks
         }
 
         [Benchmark]
-        public int Scalar()
+        public unsafe int Scalar()
         {
-            for (int i = 0; i < _array.Length; i++)
+            fixed (char* arr = _array)
             {
-                if (_array[i] == _searchValue)
+                for (int i = 0; i < Size; i++)
                 {
-                    return i;
+                    if (arr[i] == _searchValue)
+                    {
+                        return i;
+                    }
                 }
+                return -1;
             }
-            return -1;
         }
 
         [Benchmark]
@@ -72,18 +75,20 @@ namespace SveBenchmarks
                     // Compare each vector value with the target
                     Vector128<ushort> cmp = Vector128.Equals(vals, target);
 
-                    ushort cmpSum = Vector128.Sum<ushort>(cmp);
+                    // Check if there is any match in vals by doing a pairwise maximum.
+                    // The cmpMax UInt64 value will be non-zero if the character is found.
+                    ulong cmpMax = AdvSimd.Arm64.MaxPairwise(cmp, cmp).AsUInt64().ToScalar();
 
-                    if (cmpSum > 0)
+                    if (cmpMax != 0)
                     {
-                        // find index of matching item
-                        for (int j = 0; j < incr; j++)
-                        {
-                            if (cmp.GetElement(j) == ushort.MaxValue)
-                            {
-                                return i + j;
-                            }
-                        }
+                        // Convert to byte vector and extract the odd bytes into a 64-bit scalar.
+                        Vector128<byte> cmpByte = cmp.AsByte();
+                        ulong cmpUnzip = AdvSimd.Arm64.UnzipOdd(cmpByte, cmpByte).AsUInt64().ToScalar();
+
+                        // Offset is the number of trailing bits (little endian) divided by 8.
+                        int offset = BitOperations.TrailingZeroCount(cmpUnzip) >> 3;
+
+                        return i + offset;
                     }
                 }
 
@@ -110,25 +115,21 @@ namespace SveBenchmarks
                     Vector<ushort> target = new Vector<ushort>((ushort)_searchValue);
                     var pLoop = (Vector<ushort>)Sve.CreateWhileLessThanMask16Bit(i, Size);
 
-                    for (; Sve.TestFirstTrue(Sve.CreateTrueMaskUInt16(), pLoop);  i += (int)Sve.Count16BitElements())
+                    while (Sve.TestFirstTrue(Sve.CreateTrueMaskUInt16(), pLoop))
                     {
                         Vector<ushort> vals = Sve.LoadVector(pLoop, ((ushort*)arr_ptr) + i);
                         Vector<ushort> cmpVec = Sve.CompareEqual(vals, target);
 
-                        ushort cmpSum = (ushort)Sve.AddAcross(cmpVec).ToScalar();
-
-                        if (cmpSum > 0)
+                        // Test if the character is found in the current values.
+                        if (Sve.TestAnyTrue(Sve.CreateTrueMaskUInt16(), cmpVec))
                         {
-                            // find index of matching item
-                            for (int j = 0; j < Vector<ushort>.Count; j++)
-                            {
-                                if (cmpVec.GetElement(j) == 1)
-                                {
-                                    return i + j;
-                                }
-                            }
+                            // Set elements up to and including the first active element to 1 and the rest to 0.
+                            Vector<ushort> brkVec = Sve.CreateBreakAfterMask(Sve.CreateTrueMaskUInt16(), cmpVec);
+                            // The offset is the number of active elements minus 1.
+                            return (int)Sve.SaturatingIncrementByActiveElementCount(i - 1, brkVec);
                         }
 
+                        i += (int)Sve.Count16BitElements();
                         pLoop = (Vector<ushort>)Sve.CreateWhileLessThanMask16Bit(i, Size);
                     }
                 }
@@ -149,30 +150,26 @@ namespace SveBenchmarks
                     Vector<ushort> target = new Vector<ushort>((ushort)_searchValue);
                     var pLoop = (Vector<ushort>)Sve.CreateTrueMaskInt16();
 
-
-                    for (; (Size - i) > (int)Sve.Count16BitElements(); i += (int)Sve.Count16BitElements())
+                    while (i < (Size - (int)Sve.Count16BitElements()))
                     {
                         Vector<ushort> vals = Sve.LoadVector(pLoop, ((ushort*)arr_ptr) + i);
                         Vector<ushort> cmpVec = Sve.CompareEqual(vals, target);
 
-                        ushort cmpSum = (ushort)Sve.AddAcross(cmpVec).ToScalar();
-
-                        if (cmpSum > 0)
+                        // Test if the character is found in the current values.
+                        if (Sve.TestAnyTrue(Sve.CreateTrueMaskUInt16(), cmpVec))
                         {
-                            // find index of matching item
-                            for (int j = 0; j < Vector<ushort>.Count; j++)
-                            {
-                                if (cmpVec.GetElement(j) == 1)
-                                {
-                                    return i + j;
-                                }
-                            }
+                            // Set elements up to and including the first active element to 1 and the rest to 0.
+                            Vector<ushort> brkVec = Sve.CreateBreakAfterMask(Sve.CreateTrueMaskUInt16(), cmpVec);
+                            // The offset is the number of active elements minus 1.
+                            return (int)Sve.SaturatingIncrementByActiveElementCount(i - 1, brkVec);
                         }
+
+                        i += (int)Sve.Count16BitElements();
                     }
 
                     for (; i < Size; i++)
                     {
-                        if (_array[i] == _searchValue)
+                        if (arr_ptr[i] == _searchValue)
                             return i;
                     }